refactor: 서비스 내의 Double 및 double을 Float 및 float으로 변경

+++ backend/ddang/src/main/java/com/ddang/ddang/auction/application/dto/ReadAuctionDto.java
@@ -27,7 +27,7 @@ public record ReadAuctionDto(
         Long sellerId,
         Long sellerProfileId,
         String sellerName,
-        double sellerReliability,
+        Float sellerReliability,
         boolean isSellerDeleted,
         AuctionStatus auctionStatus
 ) {
