+++ backend/ddang/src/main/java/com/ddang/ddang/report/application/dto/ReadUserInReportDto.java
@@ -7,7 +7,7 @@ public record ReadUserInReportDto(
         Long id,
         String name,
         Long profileImageId,
-        double reliability,
+        Float reliability,
         String oauthId,
         boolean isSellerDeleted
 ) {
