+++ backend/ddang/src/main/java/com/ddang/ddang/review/application/dto/ReadReviewDetailDto.java
@@ -4,9 +4,9 @@ import com.ddang.ddang.review.domain.Review;
 
 import javax.annotation.Nullable;
 
-public record ReadReviewDetailDto(@Nullable Double score, @Nullable String content) {
+public record ReadReviewDetailDto(@Nullable Float score, @Nullable String content) {
 
-    private static final Double EMPTY_SCORE = null;
+    private static final Float EMPTY_SCORE = null;
     private static final String EMPTY_CONTENT = null;
     public static final ReadReviewDetailDto EMPTY = new ReadReviewDetailDto(EMPTY_SCORE, EMPTY_CONTENT);
